- Separate requests in src/sagas/backend.ts to its own file. - Change import in backend saga test to reflect changes. - Change import in academy game to reflect changes.
